Remove library target
Make cargo-sync-rdme a binary-only crate.

Drop the forwarding lib entrypoint and the docs.rs-specific release
and CI plumbing that depended on it.
