+++ release.toml
@@ -5,7 +5,5 @@ pre-release-replacements = [
   {file = "CHANGELOG.md", search = "ReleaseDate", replace = "{{date}}"},
   {file = "CHANGELOG.md", search = "<!-- next-header -->", replace = "<!-- next-header -->\n\n## [Unreleased] - ReleaseDate", exactly = 1},
   {file = "CHANGELOG.md", search = "<!-- next-url -->", replace = "<!-- next-url -->\n[Unreleased]: https://github.com/gifnksm/cargo-sync-rdme/compare/{{tag_name}}...HEAD", exactly = 1},
-  {file = "src/lib.rs", search = "^//! cargo-sync-rdme = \".*\"$", replace = "//! cargo-sync-rdme = \"{{version}}\"", exactly = 1},
-  {file = "src/lib.rs", search = "^#!\\[doc\\(html_root_url = \"https://docs.rs/cargo-sync-rdme/.*\"\\)\\]$", replace = "#![doc(html_root_url = \"https://docs.rs/cargo-sync-rdme/{{version}}\")]", exactly = 1},
 ]
 pre-release-hook = ["just", "pre-release"]
